Don't force extension options when there is a project file
If we have a `phpstan.neon` or `phpstan.neon.dist`, there's no point in the
extension forcing its own configuration onto PHPStan, as it's expected that all
the necessary configuration will be in one of those files, in order to be able
to run PHPStan with the same configuration in other contexts as well.

Without this change, it's necessary to also manage the extension's configuration
(e.g. analysis level) per-workspace, instead of having the extension just use
the configuration already present in the project.